Restored texture merging for Gex3
